requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.3.1
Flask==3.0.3
Flask-Mail==0.10.0
Flask-Session==0.8.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.20.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2026.1.post1
Werkzeug==3.0.3
